finnaly moved completely to datastore | linked xml preferences to datastore
