+++ app/build.gradle
@@ -77,9 +77,17 @@ dependencies {
     kapt "androidx.room:room-compiler:2.2.6"
 
     //Android Navigation Architecture
-    implementation 'androidx.navigation:navigation-fragment-ktx:2.3.4'
-    implementation 'androidx.navigation:navigation-ui-ktx:2.3.4'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.3.5'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.3.5'
 
     //New Material Design
     implementation 'com.google.android.material:material:1.3.0'
+
+    //Login interceptor
+    implementation("com.squareup.okhttp3:logging-interceptor:4.9.0")
+
+
+    implementation "androidx.datastore:datastore-preferences:1.0.0-alpha08"
+    implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.4.0-alpha01"
+    implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.3.1"
 }
