--- app/src/main/java/com/example/mytrashyapp/MainActivity.kt
@@ -1,76 +1,0 @@
-package com.example.mytrashyapp
-
-
-import android.content.res.Configuration
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.appcompat.app.AppCompatDelegate
-import androidx.appcompat.app.AppCompatDelegate.MODE_NIGHT_NO
-import androidx.appcompat.app.AppCompatDelegate.MODE_NIGHT_YES
-import androidx.drawerlayout.widget.DrawerLayout
-import androidx.navigation.NavController
-import androidx.navigation.findNavController
-import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.navigateUp
-import androidx.navigation.ui.setupActionBarWithNavController
-import androidx.navigation.ui.setupWithNavController
-import androidx.preference.PreferenceManager
-import com.example.mytrashyapp.databinding.ActivityMainBinding
-
-
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityMainBinding
-
-    private lateinit var navController: NavController
-    private lateinit var drawerLayout: DrawerLayout
-    private lateinit var appBarConfiguration: AppBarConfiguration
-
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setupUI()
-
-        setContentView(binding.root)
-
-        navController = findNavController(R.id.fragment)
-        drawerLayout = binding.drawerLayout
-        appBarConfiguration = AppBarConfiguration(navController.graph, drawerLayout)
-
-        binding.navigationView.setupWithNavController(navController)
-        setupActionBarWithNavController(navController, appBarConfiguration)
-    }
-
-
-    override fun onSupportNavigateUp(): Boolean {
-        val navController = findNavController(R.id.fragment)
-        return navController.navigateUp(appBarConfiguration)
-                || super.onSupportNavigateUp()
-    }
-
-    private fun setupUI() {
-        val sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this)
-        val isLightModeOn = sharedPreferences.getBoolean("uiMode", true)
-        if(isLightModeOn) {
-            AppCompatDelegate.setDefaultNightMode(MODE_NIGHT_NO)
-        } else {
-            AppCompatDelegate.setDefaultNightMode(MODE_NIGHT_YES)
-        }
-    }
-
-
-    override fun onConfigurationChanged(newConfig: Configuration) {
-        super.onConfigurationChanged(newConfig)
-
-        // Checks the orientation of the screen
-        if (newConfig.orientation === Configuration.ORIENTATION_LANDSCAPE) {
-            println("landscape")
-        } else if (newConfig.orientation === Configuration.ORIENTATION_PORTRAIT) {
-            println("portrait")
-        }
-    }
-
-
-
-}
