--- app/src/main/java/com/example/mytrashyapp/SettingsFragment.kt
@@ -1,45 +1,0 @@
-package com.example.mytrashyapp
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.appcompat.app.AppCompatDelegate
-import androidx.preference.Preference
-import androidx.preference.PreferenceFragmentCompat
-import androidx.preference.SwitchPreferenceCompat
-
-class SettingsFragment : PreferenceFragmentCompat() {
-
-    override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
-        setPreferencesFromResource(R.xml.root_preferences, rootKey)
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-
-        setLightModeListener()
-
-        return super.onCreateView(inflater, container, savedInstanceState)
-    }
-
-    fun setLightModeListener() {
-        val uiPreference = preferenceManager.findPreference<Preference>("uiMode") as SwitchPreferenceCompat
-        uiPreference.setOnPreferenceChangeListener { _, newValue ->
-            if(newValue as Boolean){
-                println("Light")
-                AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
-            }else {
-                println("Dark")
-                AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
-            }
-            true
-        }
-    }
-
-
-
-}
