+++ app/src/main/java/com/example/mytrashyapp/data/network/Resource.kt
@@ -1,0 +1,12 @@
+package com.example.mytrashyapp.data.network
+
+import okhttp3.ResponseBody
+
+sealed class Resource<out T> {
+    data class Success<out T>(val value: T): Resource<T>()
+    data class Failure(
+            val isNetworkErr: Boolean,
+            val errCode: Int?,
+            val errBody: ResponseBody?
+    ): Resource<Nothing>()
+}
