+++ app/src/main/java/com/example/mytrashyapp/data/network/repository/AuthRep.kt
@@ -1,0 +1,11 @@
+package com.example.mytrashyapp.data.network.repository
+
+import com.example.mytrashyapp.data.network.AuthApi
+
+class AuthRep(
+        private val api: AuthApi
+): BaseRep() {
+    suspend fun login(token: String) = safeApiCall {
+        api.login(token)
+    }
+}
