+++ app/src/main/java/com/example/mytrashyapp/data/network/responses/LoginResp.kt
@@ -1,0 +1,5 @@
+package com.example.mytrashyapp.data.network.responses
+
+data class LoginResp(
+    val jwt: String
+)
