--- app/src/main/java/com/example/mytrashyapp/library/screens/genres/models/Genre.kt
@@ -1,8 +1,0 @@
-package com.example.mytrashyapp.library.screens.genres.models
-
-import com.example.mytrashyapp.library.screens.songs.models.Song
-
-class Genre (
-        val name : String = "Unknown",
-        val songs : Array<Song>
-        )
