--- app/src/main/java/com/example/mytrashyapp/library/screens/songs/viewHolders/SongViewHolder.kt
@@ -1,13 +1,0 @@
-package com.example.mytrashyapp.library.screens.songs.viewHolders
-
-import androidx.recyclerview.widget.RecyclerView
-import com.example.mytrashyapp.databinding.SongBinding
-import com.example.mytrashyapp.library.screens.songs.models.Song
-
-class SongViewHolder (private val binding: SongBinding): RecyclerView.ViewHolder(binding.root) {
-    fun bind(song: Song) {
-        binding.songName.text = song.name
-        binding.songArtist.text = song.artist
-        binding.imageView.setImageResource(song.image)
-    }
-}
