--- app/src/main/java/com/example/mytrashyapp/network/Resource.kt
@@ -1,12 +1,0 @@
-package com.example.mytrashyapp.network
-
-import okhttp3.ResponseBody
-
-sealed class Resource<out T> {
-    data class Success<out T>(val value: T): Resource<T>()
-    data class Failure(
-            val isNetworkErr: Boolean,
-            val errCode: Int?,
-            val errBody: ResponseBody?
-    ): Resource<Nothing>()
-}
