--- app/src/main/java/com/example/mytrashyapp/network/repository/AuthRep.kt
@@ -1,11 +1,0 @@
-package com.example.mytrashyapp.network.repository
-
-import com.example.mytrashyapp.network.AuthApi
-
-class AuthRep(
-        private val api: AuthApi
-): BaseRep() {
-    suspend fun login(token: String) = safeApiCall {
-        api.login(token)
-    }
-}
