--- app/src/main/java/com/example/mytrashyapp/network/responses/LoginResp.kt
@@ -1,5 +1,0 @@
-package com.example.mytrashyapp.network.responses
-
-data class LoginResp(
-    val jwt: String
-)
