+++ app/src/main/java/com/example/mytrashyapp/ui/MainActivity.kt
@@ -1,0 +1,104 @@
+package com.example.mytrashyapp.ui
+
+
+import android.content.res.Configuration
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatDelegate
+import androidx.appcompat.app.AppCompatDelegate.MODE_NIGHT_NO
+import androidx.appcompat.app.AppCompatDelegate.MODE_NIGHT_YES
+import androidx.drawerlayout.widget.DrawerLayout
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentTransaction
+import androidx.lifecycle.Observer
+import androidx.lifecycle.asLiveData
+import androidx.lifecycle.lifecycleScope
+import androidx.navigation.NavController
+import androidx.navigation.findNavController
+import androidx.navigation.ui.AppBarConfiguration
+import androidx.navigation.ui.navigateUp
+import androidx.navigation.ui.setupActionBarWithNavController
+import androidx.navigation.ui.setupWithNavController
+import androidx.preference.PreferenceManager
+import com.example.mytrashyapp.R
+import com.example.mytrashyapp.data.preferences.UserPreferences
+import com.example.mytrashyapp.databinding.ActivityMainBinding
+import com.example.mytrashyapp.util.observeOnce
+import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.runBlocking
+
+
+class MainActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityMainBinding
+
+    private lateinit var navController: NavController
+    private lateinit var drawerLayout: DrawerLayout
+    private lateinit var appBarConfiguration: AppBarConfiguration
+
+    private lateinit var pref: UserPreferences
+
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+
+        pref = UserPreferences(this)
+
+        lifecycleScope.launch { //cashing the uiMode if possible
+            pref.uiMode.first()  // however the next thing is setupUI which is runBlocking so it doesnt matter that much
+        }
+
+        setupUI()
+
+
+        setContentView(binding.root)
+
+        navController = findNavController(R.id.fragment)
+        drawerLayout = binding.drawerLayout
+        appBarConfiguration = AppBarConfiguration(navController.graph, drawerLayout)
+
+        binding.navigationView.setupWithNavController(navController)
+        setupActionBarWithNavController(navController, appBarConfiguration)
+
+
+        pref.uiMode.asLiveData().observe(this, Observer {
+            println(it)
+        })
+    }
+
+
+
+    override fun onSupportNavigateUp(): Boolean {
+        val navController = findNavController(R.id.fragment)
+        return navController.navigateUp(appBarConfiguration)
+                || super.onSupportNavigateUp()
+    }
+
+    private fun setupUI() {
+        val lightMode = runBlocking { pref.uiMode.first() }
+        println("suca $lightMode")
+        if(lightMode) {
+            AppCompatDelegate.setDefaultNightMode(MODE_NIGHT_NO)
+        } else {
+            AppCompatDelegate.setDefaultNightMode(MODE_NIGHT_YES)
+        }
+
+    }
+
+
+    override fun onConfigurationChanged(newConfig: Configuration) {
+        super.onConfigurationChanged(newConfig)
+
+        // Checks the orientation of the screen
+        if (newConfig.orientation === Configuration.ORIENTATION_LANDSCAPE) {
+            println("landscape")
+        } else if (newConfig.orientation === Configuration.ORIENTATION_PORTRAIT) {
+            println("portrait")
+        }
+    }
+
+
+
+}
