+++ app/src/main/java/com/example/mytrashyapp/ui/library/screens/genres/models/Genre.kt
@@ -1,0 +1,8 @@
+package com.example.mytrashyapp.ui.library.screens.genres.models
+
+import com.example.mytrashyapp.ui.library.screens.songs.models.Song
+
+class Genre (
+        val name : String = "Unknown",
+        val songs : Array<Song>
+        )
