+++ app/src/main/java/com/example/mytrashyapp/ui/library/screens/songs/models/Song.kt
@@ -1,0 +1,9 @@
+package com.example.mytrashyapp.ui.library.screens.songs.models
+
+import android.widget.ImageView
+
+class Song (
+        val image: Int = -1,
+        val name: String,
+        val artist: String
+        )
