+++ app/src/main/res/layout/fragment_library.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".library.Library">
+    tools:context=".ui.library.Library">
 
 
     <com.google.android.material.tabs.TabLayout
