+++ app/src/main/res/layout/fragment_songs.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".library.screens.songs.SongsFragment">
+    tools:context=".ui.library.screens.songs.SongsFragment">
 
 
     <androidx.recyclerview.widget.RecyclerView
